% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statdes.R
\name{statdes}
\alias{statdes}
\title{Membuat Histogram, boxplot, dan ringkasan statistik}
\usage{
statdes(x)
}
\arguments{
\item{x}{numeric variable}
}
\description{
Membuat Histogram, boxplot, dan ringkasan statistik
}
